Split up sensors API
