+++ client/src/api/v2/sensors/index.js
@@ -1,0 +1,2 @@
+export * from './status';
+export * from './timeSeries';
