+++ client/src/api/v2/sensors/status.js
@@ -1,45 +1,4 @@
-import { useState, useCallback } from 'react';
-import { useTimeSeries } from 'utils/timeSeries';
-import { useChannel } from './socket';
-
-export function useData() {
-  const [data, setData] = useState(null);
-
-  const handler = useCallback((newData) => {
-    setData(newData);
-  }, []);
-  useChannel('data', handler);
-
-  return data;
-}
-
-export function useSensorTimeSeries(sensor, interval) {
-  const [running, setRunning] = useState(false);
-  const {
-    series,
-    max,
-    add,
-    reset,
-  } = useTimeSeries(interval, running);
-
-  const handleData = useCallback((data) => {
-    add(data[sensor]);
-  }, [add, sensor]);
-  useChannel('data', handleData);
-
-  const handleStart = useCallback(() => {
-    reset();
-    setRunning(true);
-  }, [reset]);
-  useChannel('start', handleStart);
-
-  const handleStop = useCallback(() => {
-    setRunning(false);
-  }, []);
-  useChannel('stop', handleStop);
-
-  return { series, max };
-}
+import { useData } from './data';
 
 /** Transform sensor data from an object to a list of sensors */
 function formatData(data) {
